Add ability to tag a commit
At this point, it is tagging the commit on master that is the parent of the RC to master merge
commit, which is very specific to our team's process.

Next step is to make this more generic so that it simply tags a commit based on a given SHA.  This
should be possible to accomplish by making the script CLI take an action and params specific to that
action. For instance have a 'comment' action and a 'tag' action, each taking a different set of params.